+++ routes/projectRoute.js
@@ -1,12 +1,11 @@
 const express = require("express");
-
-const User = require("../models/User");
 const { verifyToken } = require("../middlewares/authMiddleware");
 const {
   generateSheetUrl,
   validateDb,
   validateSheet,
   getProject,
+  synchronizeController,
 } = require("../controllers/projectController");
 
 const route = express.Router();
@@ -15,5 +14,6 @@ route.get("/:id", getProject);
 route.post("/validation/db", validateDb);
 route.post("/validation/sheet", validateSheet);
 route.get("/generation/sheet", verifyToken, generateSheetUrl);
+route.post("/sync", verifyToken, synchronizeController);
 
 module.exports = route;
